Reduce repeated code in network event instrumentors
There is a bunch of boilerplate code that precedes the commit calls
on SocketReadEvent and SocketWriteEvent that was repeated in the
various instrumentors.  The code was placed in a static method called
processEvent() on each of the two events.  The instrumentors were
changed to call the static methods for event processing.