Initial work to add JFR events to async sockets.
To support asynchronous socket channels, there is a private read and
write method in AsynchronousSocketChannelImpl that all reads and writes
respectively get routed through.  The class
AsynchronousSocketChannelImplInstrumentor should instrument those
methods to decorate the CompletionHandler to generate the JFR event and
then forward to the real handler.  The implementation is less than ideal
since it allocates a decorator whether the event is logged or not.

The test (TestAsynchronousSocketChannelEvents) is incomplete and the
actual check of the event generation is disabled because there is
currently a bug where AsynchronousSocketChannelImplInstrumentor is not
actually used to instrument the AsynchronousSocketChannelImpl class.  I
set a breakpoint at JVMUpcalls::onRetransform and the instrumentation is
never attempted.  When the instrumentation problem is fixed, the test
will be completed.