Changed to a single service called EventGateway.
The event delivery service was changes to a static final value to
avoid concurrency issues.  The variable EventGateway.service holds
either the JFR service implementation or an empty stub service if
the JFR module has not been loaded.  The datagram support has it's
own event types DatagramReceive and DatagramSend now.