Network JFR events sent using JFR as a service from java.base.
Each class that used to be instrumented at runtime by the jdk.jfr module
now have a static final field that is initialized from the EventService.
If there is no service a stub is installed that does nothing.  Logic that
used to be in the jdk.jfr is now in java.base where it can be more easily
viewed and checked at compile time.