Added EventServiceLookup.
Removed the public static final service field on the EventService
interface and created the EventServiceLookup class to house the service
fetching logic.  An attempt to log events before java.base is fully
loaded would previously cause the stub to be used for all logging.  Now
the stub is returned until java.base is loaded at which time an attempt
is made to find the jfr service implementation.